Initial program to analyse .uiscript files
Since .uiscript files are modified versions of XML, HTML is the closest
language to easily work with them. With a bit of CSS/JS, they can be
styled in a web view for inspection.

A tool sat on the backburner for a long time, originally using Qt
widgets, but now using QWebEngine.

The tool can be refined over time, but it'll give a helping hand in
figuring out how the game puts UI together to come up with solutions to
our outstanding issues.